requirements.txt
apache-airflow>=2.6.0
yfinance
pandera
pandas
torch
scikit-learn
numpy
mlflow==2.22.0
optuna
# --- LA COMBINAISON GAGNANTE ---
evidently==0.4.15
pydantic==1.10.13
typing_extensions>=4.6.0